// src/models/Phone.tsx
/**
 * Les propriétés pour tous les films.
 */

export interface Phone {
  id: string;
  model: string;
  constructor: string;
  os: string;
  releaseDate: number;
  salerAvatar: string;
  saler: string;
  description: string;
  salerGender: string;
  salerCity: string;
  salerCountry: string;
  phone: string;
  price: number;
}